source/requirements.txt
# Backend
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
together
python-multipart
sentence_transformers
gunicorn
psycopg2-binary

# Frontend
pandas
requests

#Framework
hinteval
